requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
django-allauth==0.61.1
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
filelock==3.13.1
idna==3.6
oauthlib==3.2.2
phonenumbers==8.13.33
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.0
